mcp_components/jira_container/requirements.txt
atlassian-python-api>=4.0.7
fastmcp>=3.1.0
httpx>=0.28.1
langchain-core>=1.2.17
langsmith>=0.7.13
python-dotenv>=1.2.2
uvicorn>=0.41.0
